測光ヒストグラムの受け皿を32ビットへ。白飛びの多い場面で暗く読んでいた
「昼が明るすぎる」の真因。自前で作る輝度ヒストグラムの受け皿が uint16 で、
1 ビンの上限が 65,535 だった。内蔵カメラの画像は 2040x1536 = 313 万画素あるので、
同じ明るさの画素が全体の 2.1% を超えるとビンが溢れて巻き戻る。平らな白い空は
これを軽く超え、明側が数えられずに中央値が暗く出ていた。
場面を実際より暗いと判断するので、露出を明るいまま残していた。

実測(2026-09-20 朝の窓辺・朝0):
  実画像の中央値 173 / 白飛び(250以上) 36%
  アプリの測光   138 相当  → 0.7 段明るく仕上がる
16 ビットで巻き戻した中央値を計算すると 138 になり、ログの値と一致した。

出方が場面次第なので今まで表に出なかった:
  サインシャンド(草原・階調が滑らか・白飛び 0〜7%) ずれ -0.4〜-3 画素
  窓辺(白飛び 29〜37%)                          ずれ +17〜+38 画素
キヤノン機はサムネイル(小さい画像)を測るので溢れない。露出の無段化とは無関係で、
以前から入っていた。

直し方:
- jpglm::lumaHistogram の受け皿を uint32 に。呼ぶ側 2 か所(内蔵カメラの測光と
  キヤノンのサムネイル測光)も uint32 へ。
- expo::histMedian に uint32 版を足す(実装は共通のテンプレート)。
  カメラが答えるヒストグラムは小さい画像の集計なので 16 ビットのまま。

実機確認(AQUOS SH-M08・白飛び 39% の窓辺):
  溢れるビン 1 個・最大 1,023,890(16 ビット上限の 15.6 倍)
  アプリの測光 x=0.8440 → 画素値 215.2 / 実画像の中央値 219  (差 4)
  同じ画像を 16 ビットで数えると 169 = 0.84 段暗い(直前までの姿)
Pixel 6 でも周期・露出とも従来どおり。

検証: 95_exposureLoop ALL PASS(受け皿の試験を追加) / 96_planCcm PASS 59 /
3 ターゲットビルド。スマホ 2 台に書き込み済み。エッジ 2 台は未接続で書き込み待ち。

+++ 40_src/20_HolyGrailEntity/10_common/src/apiCanonCCAPI.cpp
@@ -2276,7 +2276,7 @@ errCode apiCanonCCAPI::thumbMeterCore(meterResult& out, int budgetMs, const std:
 	void* tf = tool::startElapse();
 	const std::string url = base + path + "?kind=thumbnail";
 	std::string jpg;
-	uint16_t hist[256];
+	uint32_t hist[256];
 	int  w = 0, h = 0;
 	bool got = false, dec = false;
 	int  decodeMs = 0;
